Converting goals activity datatype to INT whhich will used to pull from the sport types table to set activity. Fixing bug where Thyleaf disabled automatic checking of a Sport when going to the goals page. Moving file structure to their designated folders and deleting the goals folder.
