+++ src/database/schema/008-goals_table_update.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `goals` MODIFY COLUMN activity INT;
