+++ src/main/java/com/depaul/trilog/web/GoalsController.java
@@ -1,5 +1,8 @@
-package com.depaul.trilog.goals;
+package com.depaul.trilog.web;
 
+import com.depaul.trilog.dao.GoalsRepository;
+import com.depaul.trilog.entities.Goals;
+import com.depaul.trilog.services.GoalsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
