+++ src/main/resources/templates/goals.html
@@ -19,9 +19,9 @@
 
                         <fieldset id="Activity" >
                             <label for="goalname">Activity</label>
-                            <input type="radio" name="activitySelection" id="swimmingChoice" value="swimming" checked th:field = "*{activity}" placeholder = "activity" > Swimming
-                            <input type="radio" name="activitySelection" id="runningChoice" value="running" th:field = "*{activity}" placeholder = "activity"> Running
-                            <input type="radio" name="activitySelection" id="cyclingChoice" value="cycling" th:field = "*{activity}" placeholder = "activity"> Cycling
+                            <input type="radio" name="activitySelection1" id="swimmingChoice" value=1  th:field = "*{activity}" placeholder = "activity" checked="checked"> Swimming
+                            <input type="radio" name="activitySelection2" id="runningChoice" value=3 th:field = "*{activity}" placeholder = "activity"> Running
+                            <input type="radio" name="activitySelection3" id="cyclingChoice" value=2 th:field = "*{activity}" placeholder = "activity"> Cycling
                         </fieldset>
                         <div class="slidecontainer">
                             Distance:
